GraphMatching/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
alabaster=0.7.12=pyhd3eb1b0_0
anyio=3.4.0=py38haa244fe_0
appdirs=1.4.4=pyhd3eb1b0_0
argh=0.26.2=py38_0
argon2-cffi=21.1.0=py38h294d835_2
arrow=0.13.1=py38_0
astroid=2.6.6=py38haa95532_0
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py38he774522_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py38h2a96729_1
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py38haa244fe_1
cffi=1.15.0=py38h2bbff1b_0
chardet=4.0.0=py38haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cookiecutter=1.7.2=pyhd3eb1b0_0
cryptography=35.0.0=py38h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py38hd77b12b_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.13=pypi_0
diff-match-patch=20200713=pyhd3eb1b0_0
dimod=0.10.7=pypi_0
diskcache=5.2.1=pypi_0
docutils=0.17.1=py38haa95532_1
dwave-cloud-client=0.9.2=pypi_0
dwave-drivers=0.4.4=pypi_0
dwave-greedy=0.2.1=pypi_0
dwave-hybrid=0.6.4=pypi_0
dwave-inspector=0.2.7=pypi_0
dwave-inspectorapp=0.2.3=pypi_0
dwave-neal=0.5.8=pypi_0
dwave-networkx=0.8.10=pypi_0
dwave-ocean-sdk=4.2.0=pypi_0
dwave-preprocessing=0.3.1.post0=pypi_0
dwave-qbsolv=0.3.3=pypi_0
dwave-system=1.10.0=pypi_0
dwave-tabu=0.4.2=pypi_0
dwavebinarycsp=0.1.3=pypi_0
easydict=1.9=pypi_0
entrypoints=0.3=py38_0
fasteners=0.16.3=pypi_0
flake8=3.9.2=pyhd3eb1b0_0
flask=2.0.2=pypi_0
fonttools=4.28.2=py38h294d835_0
freetype=2.10.4=h546665d_1
homebase=1.0.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.2=pyhd3eb1b0_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=4.8.1=py38haa95532_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.1=hd3eb1b0_0
inflection=0.5.1=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.4.1=py38haa95532_1
ipython=7.29.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.0=py38haa95532_1
jinja2=3.0.3=pypi_0
jinja2-time=0.2.0=pyhd3eb1b0_2
jpeg=9d=h2bbff1b_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.9.1=py38haa95532_0
jupyter_server=1.12.0=pyhd8ed1ab_0
jupyterlab=3.2.4=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.8.2=pyhd8ed1ab_0
keyring=23.1.0=py38haa95532_0
kiwisolver=1.3.2=py38hbd9d945_1
lazy-object-proxy=1.6.0=py38h2bbff1b_0
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libpng=1.6.37=h2a8f88b_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.0.10=h21b02b4_1005
libuv=1.40.0=he774522_0
libwebp=1.2.1=h57928b3_0
libwebp-base=1.2.1=h8ffe710_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=pypi_0
matplotlib=3.5.0=py38haa244fe_0
matplotlib-base=3.5.0=py38h1f000d6_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py38_1
minorminer=0.2.6=pypi_0
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
mypy_extensions=0.4.3=py38_0
nbclassic=0.3.4=pyhd8ed1ab_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pypi_0
notebook=6.4.6=pyha770c72_0
numexpr=2.7.3=py38hb80d3ca_1
numpy=1.21.2=py38hfca59bb_0
numpy-base=1.21.2=py38h0829f74_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1l=h8ffe710_0
ortools=8.2.8710=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py38h6214cd6_0
pandocfilters=1.4.3=py38haa95532_1
paramiko=2.7.2=py_0
parso=0.8.2=pyhd3eb1b0_0
pathspec=0.7.0=py_0
penaltymodel=0.16.5=pypi_0
penaltymodel-cache=0.4.4=pypi_0
penaltymodel-lp=0.1.5=pypi_0
penaltymodel-mip=0.2.5=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38hd45dc43_0
pip=21.0.1=py38haa95532_0
plucky=0.4.3=pypi_0
pluggy=1.0.0=py38haa95532_0
plyfile=0.7.4=pyhd8ed1ab_0
poyo=0.5.0=pyhd3eb1b0_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=pypi_0
psutil=5.8.0=py38h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pygmtools=0.1.17=pypi_0
pylint=2.9.6=py38haa95532_1
pyls-spyder=0.4.0=pyhd3eb1b0_0
pynacl=1.4.0=py38h62dcd97_1
pynvml=11.4.1=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py38ha925a31_4
pyqubo=1.0.13=pypi_0
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
python=3.8.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lsp-black=1.0.0=pyhd3eb1b0_0
python-lsp-jsonrpc=1.0.0=pyhd3eb1b0_0
python-lsp-server=1.2.4=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.10.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywin32=228=py38hbaba5e8_1
pywin32-ctypes=0.2.0=py38_1000
pywinpty=0.5.7=py38_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=22.2.1=py38hd77b12b_1
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.3=pyhd3eb1b0_0
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
regex=2021.8.3=py38h2bbff1b_0
requests=2.26.0=pyhd3eb1b0_0
rope=0.21.1=pyhd3eb1b0_0
rtree=0.9.7=py38h2eaa2aa_1
scipy=1.7.1=py38hbe87c03_2
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py38haa244fe_2
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sphinx=4.2.0=pyhd3eb1b0_1
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=5.1.5=py38haa95532_1
spyder-kernels=2.1.3=py38haa95532_0
sqlite=3.36.0=h2bbff1b_0
tensorboardx=2.4.1=pypi_0
terminado=0.9.4=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tinycss=0.4=pyhd3eb1b0_1002
tk=8.6.11=h8ffe710_1
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typed-ast=1.4.3=py38h2bbff1b_1
typing_extensions=3.10.0.2=pyh06a4308_0
ujson=4.0.2=py38hd77b12b_0
unicodedata2=13.0.0.post2=py38h294d835_4
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.7=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.3=py38haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.2.1=py38haa244fe_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
whichcraft=0.6.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
wrapt=1.12.1=py38he774522_1
xlrd=2.0.1=pypi_0
xlwt=1.3.0=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h6255e5f_0
